Move `registries` singletons into `src/stores` folder
- create `src/stores` folder
- deprecate notice for `core/catalog/cataloglayersstoresregistry`
- deprecate notice for `core/editing/sessionsregistry`
- deprecate notice for `core/map/maplayersstoresregistry`
- deprecate notice for `core/plugin/pluginsregistry`
- deprecate notice for `core/project/projectsregistry`
- deprecate notice for `gui/component/componentsregistry`
- deprecate notice for `gui/map/control/registry`
- consolidate singletons definitions: `LayersStoresRegistry`,` ComponentsRegistry`, `ControlsRegistry`,` MapLayersStoresRegistry`, `PluginsRegistry`,` ProjectsRegistry`, `SessionsRegistry` within a single folder
