+++ src/app/gui/component/componentsregistry.js
@@ -1,36 +1,12 @@
-const G3WObject = require('core/g3wobject');
-const {base, inherit} = require('core/utils/utils');
+/**
+ * DEPRECATED: this file will be removed after v3.4 (use "stores/components.js" instead)
+ */
+import ComponentsRegistry from 'stores/components';
 
-//class Componet Registry (singleton)
-// store all components added
-function ComponentsRegistry() {
-  this.components = {};
-  this.registerComponent = function(component) {
-    const id = component.getId();
-    if (!this.components[id]) {
-      this.components[id] = component;
-      this.emit('componentregistered', component);
-    }
-  };
+/**
+ * FIXME: application is broken using like the following line
+ */
+// import { ComponentsRegistry } from 'stores';
 
-  this.getComponent = function(id) {
-    return this.components[id];
-  };
-
-  this.getComponents = function() {
-    return this.components;
-  };
-
-  this.unregisterComponent = function(id) {
-    const component = this.components[id];
-    if (component) {
-      if (typeof component.destroy === 'function') component.destroy();
-      this.components[id] = null;
-    }
-    return component;
-  };
-  base(this);
-}
-inherit(ComponentsRegistry, G3WObject);
-
-module.exports = new ComponentsRegistry;
+ module.exports = ComponentsRegistry;
+
