+++ src/stores/components.js
@@ -1,0 +1,41 @@
+/**
+ * ORIGINAL SOURCE: src/app/gui/component/componentsregistry.js@v3.4
+ */
+
+const G3WObject = require('core/g3wobject');
+const {base, inherit} = require('core/utils/utils');
+
+//class Componet Registry (singleton)
+// store all components added
+function ComponentsRegistry() {
+  this.components = {};
+  this.registerComponent = function(component) {
+    const id = component.getId();
+    if (!this.components[id]) {
+      this.components[id] = component;
+      this.emit('componentregistered', component);
+    }
+  };
+
+  this.getComponent = function(id) {
+    return this.components[id];
+  };
+
+  this.getComponents = function() {
+    return this.components;
+  };
+
+  this.unregisterComponent = function(id) {
+    const component = this.components[id];
+    if (component) {
+      if (typeof component.destroy === 'function') component.destroy();
+      this.components[id] = null;
+    }
+    return component;
+  };
+  base(this);
+}
+
+inherit(ComponentsRegistry, G3WObject);
+
+export default new ComponentsRegistry();
